+++ src/mod_ex_flow_utils/mod.rs
@@ -1,5 +1,5 @@
-pub mod datatype_ex_flow;
-pub mod entities;
+
+
 pub mod errors;
-pub mod uri;
-pub mod utils_ex_flow;
+pub mod uri_endpoints;
+pub mod utils;
