+++ src/mod_ex_flow_utils/uri_endpoints.rs
@@ -1,6 +1,13 @@
+///
+/// For ExFlow Service
+///
 pub const EX_FLOW_SERVICE_API_SCOPE: &'static str = "/api/v1";
 pub const EX_FLOW_SERVICE_API_IR_REGISTER: &'static str = "/ir/register";
 
+
+///
+/// For ExFlow Runtime
+///
 pub const EX_FLOW_RUNTIME_API_SCOPE: &'static str = "/api/v1";
 pub const EX_FLOW_RUNTIME_API_RUN_PIPELINE: &'static str = "/run_pipeline";
 pub const EX_FLOW_RUNTIME_API_GET_PIPELINE: &'static str = "/get_pipeline";
