+++ src/mod_service_api/entities.rs
@@ -1,6 +1,5 @@
 use serde::{Deserialize, Serialize};
-
-use crate::mod_ex_flow_utils::entities::SystemInformation;
+use crate::mod_ex_flow_utils::utils::SystemInformation;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ExFlowRuntimeRegisterRequest {
