+++ src/mod_service_cli/service_cli.rs
@@ -4,8 +4,8 @@ use actix_web_opentelemetry::RequestTracing;
 use clap::Parser;
 use log::info;
 
-use crate::mod_ex_flow_utils::uri::{EX_FLOW_SERVICE_API_IR_REGISTER, EX_FLOW_SERVICE_API_SCOPE};
-use crate::mod_ex_flow_utils::utils_ex_flow::set_global_apm_tracing;
+use crate::mod_ex_flow_utils::uri_endpoints::{EX_FLOW_SERVICE_API_IR_REGISTER, EX_FLOW_SERVICE_API_SCOPE};
+use crate::mod_ex_flow_utils::utils::set_global_apm_tracing;
 use crate::mod_service_api::service_api::post_register_runtime;
 
 const SERVICE_NAME: &'static str = "ExFlow-Service";
